Use Async Task to prevent an ANR
In the Launcher application, the wallpaper chooser activity uses
the main UI thread to make a binder call to the WallPaper Manager
service. This happens when the user selects a wall paper and sets
it. However, an ANR occurs when the binder call does not respond
in time. Hence, such calls should not be done in the main UI
thread.

Change-Id: I3bcba39f3086b7d1842b4d1c2f1ff1f71786041c