Launcher: Fix delay responding when dragging an icon with two fingers
This issue is because Launcher app do not catch ACTION_POINTER_DOWN and
ACTION_POINTER_UP events, so it does not move the DragView to the second
finger until there is a ACTION_MOVE event.

Add ACTION_POINTER_DOWN and ACTION_POINTER_UP event handling to OnTouchEvent
method will fix this issue.

Change-Id: Ibf755ca4ab128f949fb628758857a7181ed0f29c
Author: Jian Li <jian.d.li@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Beare, Bruce J <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 18432