Correct the smallest cell size for calulating widget span
Use the actual smallest cell width and height in portrait and landscape
mode when calculating the number of cells needed by a widget. Before
this fix the smallest cell dimension in the current orientation was used
to determine the cell size. This caused some strange behaviour like a
widget could get different number of cells depending on if it was added
in landscape or portrait mode since the size of the smallest cell
dimension differed for the different orientations.

Change-Id: I338cae0f1482cb7fd61251983f92f8f432348f1a