Fix for mSavedStates being out of sync
mSavedStates contains the state that should be restored on resume. Apparently the method restoreInstanceStateForRemainingPages gets called without mSavedStates being up-to-date.
This problem leads in certain circumstances to widgets being restored to an old state, although newer data was displayed before. The issue is described in the following bug report:
https://code.google.com/p/android/issues/detail?id=42534
The problem can be reproduced on the latest Nexus factory images with the procedure described in the bug report (I'll update it with information about this fix later).
With this fix being submitted, the saved state will always be up-to-date and widgets will be restored correctly.

Change-Id: Ic8be85426116f7a27585cc22b8ee46f58a785925
Signed-off-by: Michael Diener <mdienersoftware@gmail.com>