Eliminate 1 frame drawing delay at start of scroll
When ACTION_DOWN for Homescreen panel scroll occurs on a child (shortcut
or widget), start of scroll is detected via intercepted touch events in
PagedView.onInterceptTouchEvent() once an ACTION_MOVE touch event is
received which shows initial movement beyond a threshold. However, the
panel scroll position was not changed as a result of this event.
The panel scroll position did not change until the next frame, when
the next ACTION_MOVE event was delivered to PagedView.onTouchEvent().
As a result, the first draw to show any scrolling was delayed by 1
frame or 17ms.

Modified to change the scroll position as soon as the scroll start
is detected.

Change-Id: Ib76763313c37007314e3f58dfdf1ffd3468e6b3a